appBack/requirements.txt
fpdf
djangorestframework
django-cors-headers
python-telegram-bot
django-multiselectfield
telebot
tabulate
schedule